Move `Frontend_Assets` to WooCommerce integration
