+++ tests/integration/wp-includes/class-bh-wc-bitcoin-gateway-integration-Test.php
@@ -8,7 +8,7 @@
 
 namespace BrianHenryIE\WP_Bitcoin_Gateway;
 
-use BrianHenryIE\WP_Bitcoin_Gateway\Frontend\Frontend_Assets;
+use BrianHenryIE\WP_Bitcoin_Gateway\Integrations\WooCommerce\Frontend_Assets;
 use BrianHenryIE\WP_Bitcoin_Gateway\WP_Includes\I18n;
 use lucatume\WPBrowser\TestCase\WPTestCase;
