+++ tests/integration/wp-includes/class-rest-integration-Test.php
@@ -7,8 +7,6 @@
 
 namespace BrianHenryIE\WP_Bitcoin_Gateway;
 
-use BrianHenryIE\WP_Bitcoin_Gateway\Frontend\Frontend_Assets;
-use BrianHenryIE\WP_Bitcoin_Gateway\WP_Includes\I18n;
 use lucatume\WPBrowser\TestCase\WPTestCase;
 
 class Rest_Integration_Test extends WPTestCase {
